input from 'email' prints in console
